ai-agent/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pymongo
openai